VBliss_Blog/.aws-sam/cache/3bd4c421-753e-43a2-9ae8-97e95e8c9623/requirements.txt
boto3
Pillow
sagemaker
setuptools